requirements.txt
telethon==1.36.0
python-dotenv==0.19.2